+++ lib/truffle/truffle/cext.rb
@@ -2375,7 +2375,16 @@ module Truffle::CExt
   end
 
   def rb_error_frozen_object(object)
-    raise FrozenError.new("can't modify frozen #{Primitive.class(object)}", receiver: object)
+    string = nil
+    recursion = Truffle::ThreadOperations.detect_recursion object do
+      string = object.inspect
+    end
+
+    if recursion
+      string = ' ...'
+    end
+
+    raise FrozenError.new("can't modify frozen #{Primitive.class(object)}: #{string}", receiver: object)
   end
 
   def rb_tr_warn(message)